+++ security/pkg/server/ca/server_test.go
@@ -102,7 +102,7 @@ func (authz *mockAuthorizer) authorize(requester *authenticate.Caller, requested
 
 func TestCreateCertificate(t *testing.T) {
 	testCases := map[string]struct {
-		authenticators []authenticator
+		authenticators []authenticate.Authenticator
 		authorizer     *mockAuthorizer
 		ca             CertificateAuthority
 		certChain      []string
@@ -115,7 +115,7 @@ func TestCreateCertificate(t *testing.T) {
 			ca:             &mockca.FakeCA{},
 		},
 		"Unauthenticated request": {
-			authenticators: []authenticator{&mockAuthenticator{
+			authenticators: []authenticate.Authenticator{&mockAuthenticator{
 				errMsg: "Not authorized",
 			}},
 			code:       codes.Unauthenticated,
@@ -124,30 +124,30 @@ func TestCreateCertificate(t *testing.T) {
 		},
 		"CA not ready": {
 			authorizer:     &mockAuthorizer{},
-			authenticators: []authenticator{&mockAuthenticator{}},
+			authenticators: []authenticate.Authenticator{&mockAuthenticator{}},
 			ca:             &mockca.FakeCA{SignErr: caerror.NewError(caerror.CANotReady, fmt.Errorf("cannot sign"))},
 			code:           codes.Internal,
 		},
 		"Invalid CSR": {
 			authorizer:     &mockAuthorizer{},
-			authenticators: []authenticator{&mockAuthenticator{}},
+			authenticators: []authenticate.Authenticator{&mockAuthenticator{}},
 			ca:             &mockca.FakeCA{SignErr: caerror.NewError(caerror.CSRError, fmt.Errorf("cannot sign"))},
 			code:           codes.InvalidArgument,
 		},
 		"Invalid TTL": {
 			authorizer:     &mockAuthorizer{},
-			authenticators: []authenticator{&mockAuthenticator{}},
+			authenticators: []authenticate.Authenticator{&mockAuthenticator{}},
 			ca:             &mockca.FakeCA{SignErr: caerror.NewError(caerror.TTLError, fmt.Errorf("cannot sign"))},
 			code:           codes.InvalidArgument,
 		},
 		"Failed to sign": {
 			authorizer:     &mockAuthorizer{},
-			authenticators: []authenticator{&mockAuthenticator{}},
+			authenticators: []authenticate.Authenticator{&mockAuthenticator{}},
 			ca:             &mockca.FakeCA{SignErr: caerror.NewError(caerror.CertGenError, fmt.Errorf("cannot sign"))},
 			code:           codes.Internal,
 		},
 		"Successful signing": {
-			authenticators: []authenticator{&mockAuthenticator{}},
+			authenticators: []authenticate.Authenticator{&mockAuthenticator{}},
 			authorizer:     &mockAuthorizer{},
 			ca: &mockca.FakeCA{
 				SignedCert: []byte("cert"),
@@ -195,7 +195,7 @@ func TestCreateCertificate(t *testing.T) {
 
 func TestHandleCSR(t *testing.T) {
 	testCases := map[string]struct {
-		authenticators []authenticator
+		authenticators []authenticate.Authenticator
 		authorizer     *mockAuthorizer
 		ca             *mockca.FakeCA
 		csr            string
@@ -211,7 +211,7 @@ func TestHandleCSR(t *testing.T) {
 			code:           codes.Unauthenticated,
 		},
 		"Unauthenticated request": {
-			authenticators: []authenticator{&mockAuthenticator{
+			authenticators: []authenticate.Authenticator{&mockAuthenticator{
 				errMsg: "Not authorized",
 			}},
 			authorizer: &mockAuthorizer{},
@@ -220,30 +220,30 @@ func TestHandleCSR(t *testing.T) {
 		},
 		"No caller authenticated": {
 			authorizer:     &mockAuthorizer{},
-			authenticators: []authenticator{&mockAuthenticator{}},
+			authenticators: []authenticate.Authenticator{&mockAuthenticator{}},
 			code:           codes.Unauthenticated,
 		},
 		"Corrupted CSR": {
 			authorizer:     &mockAuthorizer{},
-			authenticators: []authenticator{&mockAuthenticator{identities: []string{"test"}}},
+			authenticators: []authenticate.Authenticator{&mockAuthenticator{identities: []string{"test"}}},
 			csr:            "deadbeef",
 			code:           codes.InvalidArgument,
 		},
 		"Invalid SAN CSR": {
 			authorizer:     &mockAuthorizer{},
-			authenticators: []authenticator{&mockAuthenticator{identities: []string{"test"}}},
+			authenticators: []authenticate.Authenticator{&mockAuthenticator{identities: []string{"test"}}},
 			csr:            badSanCsr,
 			code:           codes.InvalidArgument,
 		},
 		"Failed to sign": {
 			authorizer:     &mockAuthorizer{},
-			authenticators: []authenticator{&mockAuthenticator{identities: []string{"test"}}},
+			authenticators: []authenticate.Authenticator{&mockAuthenticator{identities: []string{"test"}}},
 			ca:             &mockca.FakeCA{SignErr: caerror.NewError(caerror.CANotReady, fmt.Errorf("cannot sign"))},
 			csr:            csr,
 			code:           codes.Internal,
 		},
 		"Successful signing": {
-			authenticators: []authenticator{&mockAuthenticator{identities: []string{"test"}}},
+			authenticators: []authenticate.Authenticator{&mockAuthenticator{identities: []string{"test"}}},
 			authorizer:     &mockAuthorizer{},
 			ca: &mockca.FakeCA{
 				SignedCert:    []byte("generated cert"),
@@ -256,7 +256,7 @@ func TestHandleCSR(t *testing.T) {
 			code:        codes.OK,
 		},
 		"Multiple identities received by CA signer": {
-			authenticators: []authenticator{&mockAuthenticator{identities: []string{"test1", "test2"}}},
+			authenticators: []authenticate.Authenticator{&mockAuthenticator{identities: []string{"test1", "test2"}}},
 			authorizer:     &mockAuthorizer{},
 			ca: &mockca.FakeCA{
 				SignedCert:    []byte("generated cert"),
